Fix CUDA with LLVM9
The llvm.nvvm.atomic.load.add intrinsics were removed and atomicrmw is
used in its place. The auto-upgrade of IR replaces any of the intrinsics
with atomicrmw, which is a problem when sending the optimized IR to
NVVM, which doesn't accept atomicrmw.

This commit adds patches to revert the removal of these intrinsics and
corresponding auto-upgrade of the IR.