workaround potentially broken LLVM CMake setup
With LLVM 10 the following CMake line no longer appears to work.

```
llvm_map_components_to_libnames(llvm_libs all)
```

... and this commit is a suitable workaround.

Current working hypothesis is that this might have been caused by the
change in:

https://github.com/llvm/llvm-project/commit/ab411801b82783eb7f652701ccfce81b16cf1811#diff-cebfde545aa260d20ca6b0cdc3da08daR270

However the docs at https://llvm.org/docs/CMake.html are not very clear
about this. And in addition the output of `llvm-config --components`
does list `all` as an option.

Many thanks to @Angloyna for debugging this!!