LLVM 10 (#599)
* Use std::make_unique on LLVM 10

LLVM 10 removes llvm::make_unique in favor of std::make_unique.
However, this requires C++14 and is therefore unsuitable for LLVM 9
that forces -std=c++11.  Update the code to use both conditionally.
This fixes all issues with LLVM 10.

* grab updated SVML patch and new fastmath patch

https://github.com/conda-forge/llvmdev-feedstock/blob/c706309/recipe/patches/intel-D47188-svml-VF.patch

https://github.com/conda-forge/llvmdev-feedstock/blob/c706309/recipe/patches/expect-fastmath-entrypoints-in-add-TLI-mappings.ll.patch

* updte to LLVM 10 in recipe and tests

As title

* update documented version numbers in README.rst

As title.

* update appveyor.yml comment to include LLVM 10

As title.

* update to llvmdev 10.0* on Azure pipelines

As title.

* update to llvmdev 10.0 in buildscripts

As title

* update llvmdev_manylinux1 conda-recipe

As title

* update Sphinx documentation for correct LLVM version

As title.

* temporarily set the VERSION_SUFFIX for CI

This is needed as the llvmlite package we are building needs to
differentiate itself somehow from the 'other' dev builds.

* specify the correct compiler in the build requirements

* workaround potentially broken LLVM CMake setup

With LLVM 10 the following CMake line no longer appears to work.

```
llvm_map_components_to_libnames(llvm_libs all)
```

... and this commit is a suitable workaround.

Current working hypothesis is that this might have been caused by the
change in:

https://github.com/llvm/llvm-project/commit/ab411801b82783eb7f652701ccfce81b16cf1811#diff-cebfde545aa260d20ca6b0cdc3da08daR270

However the docs at https://llvm.org/docs/CMake.html are not very clear
about this. And in addition the output of `llvm-config --components`
does list `all` as an option.

Many thanks to @Angloyna for debugging this!!

* update the comment and links for the svml patch

* add a note about broken LLVM cmake setup

As title

* update compatability matrix

As title.

* install cmake on armv7l too

* remove TODO

* bump to LLVM 10.0.1

As title.

* Make sure llvmlite build is unique.

As title.

* remove outdated comment

As title.

* implement LLVM version split

The next llvmlite release will have a split LLVM dependency. The
`aarch64` target will use LLVM 9 and all other targets will use LLVM 10.

* Revert "Fix CUDA with LLVM9"

This reverts commit 3f6612955a8458e7caf4364dc3a57863b1f93c1c.

Once https://github.com/numba/numba/pull/6030 is merged, it will no
longer be necessary to disable the autoupgrade of atomic intrinsics for
NVPTX, because LLVM from llvmlite will not be used to optimize the IR
before sending it to NVVM.

* revert LLVM documentation link to 10.0.0

The page at:

https://releases.llvm.org/10.0.1/docs/

Currently (Tue Aug  4 10:48:42 2020 GMT+2) returns a 404 - page not
found and this breaks our build as Sphinx can't setup the intersphinx
links.

* accept LLVM 9.0 once again

As we are doing a split release for 0.34.0 and `aarch64` will need LLVM
9.0.* -- we allow to build with this version again.

* Apply suggestions from code review

Fix missing word and missing punctuation.

Co-authored-by: stuartarchibald <stuartarchibald@users.noreply.github.com>

* adding link to LLVM bug report

* adding link to potential LLVM Cmake bug

As title.

* stop testing older LLVMs

We only support 10.0.* and 9.0.* from now on.

* getting ready to release 0.34.0 remove VERSION_SUFFIX

Final builds should not have a version suffix.

Co-authored-by: Michał Górny <mgorny@gentoo.org>
Co-authored-by: Graham Markall <graham@big-grey.co.uk>
Co-authored-by: stuartarchibald <stuartarchibald@users.noreply.github.com>