Disallow alwaysinline and noinline on functions
Adding "alwaysinline" and "noinline" attributes to a function results in
an error similar to the following when compiling with LLVM:

```
Attributes 'noinline and alwaysinline' are incompatible!
LLVM ERROR: Broken module found, compilation aborted!
```

and then the interpreter is killed. This commit prevents both attributes
being added at the Python level, to prevent the above error message and
termination.