Add YouCompleteMe configuration file and ignore vim swap files
YouCompleteMe (https://github.com/ycm-core/YouCompleteMe) is a
code-completion engine for vim. This PR adds a configuration file for it
that enables it to understand the C++ FFI code in llvmlite by pointing
to the correct include dirs (as long as one is working within a Conda
environment), setting the relevant C++ standard, and ensuring that all
header files are treated as C++.

Additionally, Vim swap files are added to the gitignore, so that they
don't show up in `git status`.