[llvmlite] Support opaque pointers.
LLVM 14 introduces optional opaque pointers. If opaque pointers are
enabled then there is no way to determine the pointee type. LLVM 16
makes this the default.

In llvmlite and numba this is used to determine ABI size and alignment of
types by creating a global variable and then asking for its pointee type in
order to get a TypeRef. For global variables specifically this can be shortcut
via gv->getValueType(), and doing so removes the current issues when
opaque pointers are enabled.

This change has been tested with LLVM 11-14 and so also includes some
other changes required to make llvmlite compile on those versions.