manually set the RTD theme
As of configuration file V2 for Read The Docs (RTD) the
`sphinx-rtd-theme` is no longer injected automagically by RTD. So we can
do away with all the redundant configuration and checking if we are on
RTD or not. The is now a single way to build documentation, on RTD or
locally.