Explicitly add setuptools to the conda recipe and allow Windows to use it
This is needed for building via `setup.py`. The assumption is that this
was always needed but `setuptools` had been installed by default so
llvmlite just built "by accident". With 3.12 we discovered that this
assumption no longer holds and so we must add `setuptools` to the build
requirements explicitly. In addition the conda-recipe was using `python
-s` to run the `setup.py` script which implies that packages from
`site-packages` will not be loaded. This must therefore also be removed
such that Windows can even find the installed `setuptools`.