Fix symbol visiblility for compiler-rt builtins on MacOS
The symbols in the compiler-rt builtins library are private external by
default on MacOS. We need them to be visible so we can link against them
(e.g. when building libllvmlite.so). This patch makes the symbols
visible when the CMake option `COMPILER_RT_BUILTINS_HIDE_SYMBOLS` is
off. Note that this is an existing option for the compiler-rt build, and
the flags requiring modification in this patch appear to be an
oversight.

References:

- https://discourse.llvm.org/t/lld-automatically-hide-symbols-with-prefix/73192
- https://github.com/numba/llvmlite/pull/986#issuecomment-1703872099