Fix compiler-rt atomics build
This is a backport of:

```
commit 7939ce39dac0078fef7183d6198598b99c652c88
Author: Jon Roelofs <jonathan_roelofs@apple.com>
Date:   Fri Nov 17 14:21:57 2023 -0800

    [builtins] Move cfi start's after the symbol name [NFC]

    ... in preparation for diagnosing improperly nested .cfi regions.

    See https://reviews.llvm.org/D155245
```

The ordering of `.cfi_startproc` and the function name in compiler-rt's
`assembly.h` triggers a new diagnostic about improperly nested `.cfi`
regions in Clang 18 onwards. This is fixed in compiler-rt upstream, but
needs to be applied to compiler-rt from LLVM 15 because we're building
it with Clang 18 in some cases.

References to discussions of the issue:

- https://github.com/rust-lang/llvm-project/commit/823883116fdee662a3e150a88e6aad583036ccd3
- https://issues.chromium.org/issues/40945072