Build compiler-rt builtins
This enables the build of compiler-rt builtins with the following
changes:

- Add `compiler-rt` to the list of LLVM projects to build
- Disable the build of all `compiler-rt` components except for the
  builtins. Whilst it would be interesting to add these, they seem to
  require clang to be built at the same time.
- Setting the `COMPILER_RT_BUILTINS_HIDE_SYMBOLS` CMake option to `OFF`,
  to ensure that we can link against those symbols in the general case.

Note that these changes are simpler than previous efforts towards
enabling `compiler-rt` builtins (e.g. #979) - this is because we now
build as part of the whole `llvm-project` tree, which simplifies things
and removes issues that had to be worked around when building it
separately.