Fix the metadata deduplication
The llvmlite debug metadata cache does not work as expected due to
the encoding value using address of DIToken instance and serves as
part of the key.

Solutions could be either singleton the DIToken class, or extract
the string value from the encoding object.

Since the DIToken class is designed to be a wrapper of enumeration
value, e.g. the DW_* enumerations, that should appear bare in the
emitted metadata, in this change, singleton DIToken is made to fix
the problem.

This solves llvmlit issue #1164
https://github.com/numba/llvmlite/issues/1164