Correct LTO flag use and add as option.
This patch adds LTO as an option for Unix systems, the default being
"on". It also corrects the former use whereby it was only set at link
time. LTO flags need setting at compile time to ensure that the LTO
information is present in the object files such that a link with the
LTO flag set has the information available to do LTO.