Update documentation RE LLVM_CONFIG and flags.
This updates the documentation with a note to describe that there
has been a change in the build system and the use of the
LLVM_CONFIG environment variable will need replacing with
CMAKE_PREFIX_PATH if necessary. It also notes the a couple of
build flags are actually for all Unix based systems, not just
Linux.