ffi/CMakeLists.txt: static LLVM link w -Bsymbolic
Specify -Bsymbolic linker option when
conducting LLVM-statically-linking llvmlite build.
This will thus affect the numba::llvmlite packages
but not the conda-forge::llvmlite ones, e.g.

(NOTE: We constrain the change to the "static" build variant
as this seems the one to be built and tested by the numba/llmvlite CI.)

Co-authored-by: stellaraccident