InlineAsm: fix formatting
Since a15799e7618068b8c4663c511168ecacfffb1e7e, calls to `InlineAsm`
instructions insert a newline between the constraints and the arguments.
I believe change this to have been accidentally introduced, with the
`end=''` argument to `print` in `InlineAsm` having not been noticed.

This makes the generated code harder to read and less idiomatic. For
example, for `test_inline_assembly`:

```
define void @"foo"()
{
.2:
  %".3" = call i32 asm sideeffect "mov $1, $2", "=r,r"
(i32 123)
  ret void
}
```

With this change, the code now reads as:

```
define void @"foo"()
{
.2:
  %".3" = call i32 asm sideeffect "mov $1, $2", "=r,r"(i32 123)
  ret void
}
```

I have updated the regex in `test_inline_assembly` to disallow
whitespace between the constraints and the arguments. This is in line
with other LLVM IR emitters, as seen in examples in the LLVM Language
Reference manual. For example:

```
%X = call i32 asm "bswap $0", "=r,r"(i32 %Y)
```

from https://llvm.org/docs/LangRef.html#inline-assembler-expressions.