Improve error handling on FFI binding error
This patch aims to improve the error handling when the FFI binding
library cannot be loaded or is failing on symbol lookup. By the
nature of the lazy loading in the FFI binding, if there is a
problem with the library/symbol lookup it occurs on first use. This
has manifested historically as a KeyError (because of the cache miss
in the lazy load) with a large traceback, however the issue is not
a KeyError, it's something more fundamental. This patch changes the
exception handling such that the lazy loading part now raises
exceptions with improved error messages raised from 'None' so as to
give the user a hint as to what has gone wrong and to remove the noise
from the KeyError.